requirements.txt
Django>=5.1.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-allauth>=0.57.0
Pillow>=10.0.0
python-decouple>=3.8
requests>=2.31.0
twilio>=8.0.0
google-auth>=2.20.0
facebook-sdk>=3.1.0
